Move some things, added some other things, still not working though.
Renamed ia to ai. Made no change in it.
Pieces are growing a bit, they can kind of know where they can move. Well, only pawns do, the Queen is a bit late.
ComputerPlayerUnits (CPUs) and Humans entered the game. They are useless, especially CPUs as I'm focusing on Humans to make it playable and "copy" the behaviour on the CPUs.
The Board and the Game changed a bit, mostly "graphically" speaking.